Reorganize a few files in ado_base and make all but withdrawable_tokens pub(crate)
