+++ packages/ado_base/src/auth.rs
@@ -1,0 +1,32 @@
+use crate::ADOContract;
+use common::error::ContractError;
+use cosmwasm_std::Storage;
+
+impl<'a> ADOContract<'a> {
+    /// Helper function to query if a given address is a operator.
+    ///
+    /// Returns a boolean value indicating if the given address is a operator.
+    pub fn is_operator(&self, storage: &dyn Storage, addr: &str) -> bool {
+        self.operators.has(storage, addr)
+    }
+
+    /// Helper function to query if a given address is the current contract owner.
+    ///
+    /// Returns a boolean value indicating if the given address is the contract owner.
+    pub fn is_contract_owner(
+        &self,
+        storage: &dyn Storage,
+        addr: &str,
+    ) -> Result<bool, ContractError> {
+        let owner = self.owner.load(storage)?;
+        Ok(addr == owner)
+    }
+
+    pub fn is_owner_or_operator(
+        &self,
+        storage: &dyn Storage,
+        addr: &str,
+    ) -> Result<bool, ContractError> {
+        Ok(self.is_contract_owner(storage, addr)? || self.is_operator(storage, addr))
+    }
+}
