+++ packages/ado_base/src/modules/mod.rs
@@ -1,16 +1,17 @@
 use std::convert::TryInto;
 
 use crate::state::ADOContract;
-use cosmwasm_std::{Api, DepsMut, MessageInfo, Order, QuerierWrapper, Response, Storage, Uint64};
+use cosmwasm_std::{Api, Order, QuerierWrapper, Response, Storage, Uint64};
 use cw_storage_plus::Bound;
 
-use common::{ado_base::modules::Module, error::ContractError, require};
+use common::{ado_base::modules::Module, error::ContractError};
 
+pub mod execute;
 pub mod hooks;
 pub mod query;
 
 impl<'a> ADOContract<'a> {
-    pub fn register_modules(
+    pub(crate) fn register_modules(
         &self,
         sender: &str,
         storage: &mut dyn Storage,
@@ -25,70 +26,6 @@ impl<'a> ADOContract<'a> {
                 .add_submessages(register_response.messages)
         }
         Ok(resp)
-    }
-
-    /// A wrapper for `fn register_module`. The parameters are "extracted" from `DepsMut` to be able to
-    /// execute this in a loop without cloning.
-    pub(crate) fn execute_register_module(
-        &self,
-        storage: &mut dyn Storage,
-        sender: &str,
-        module: Module,
-        should_validate: bool,
-    ) -> Result<Response, ContractError> {
-        require(
-            self.is_owner_or_operator(storage, sender)?,
-            ContractError::Unauthorized {},
-        )?;
-        let resp = Response::default();
-        let idx = self.register_module(storage, &module)?;
-        if should_validate {
-            self.validate_modules(&self.load_modules(storage)?, &self.ado_type.load(storage)?)?;
-        }
-        Ok(resp
-            .add_attribute("action", "register_module")
-            .add_attribute("module_idx", idx.to_string()))
-    }
-
-    /// A wrapper for `fn alter_module`.
-    pub(crate) fn execute_alter_module(
-        &self,
-        deps: DepsMut,
-        info: MessageInfo,
-        module_idx: Uint64,
-        module: Module,
-    ) -> Result<Response, ContractError> {
-        let addr = info.sender.as_str();
-        require(
-            self.is_owner_or_operator(deps.storage, addr)?,
-            ContractError::Unauthorized {},
-        )?;
-        self.alter_module(deps.storage, module_idx, &module)?;
-        self.validate_modules(
-            &self.load_modules(deps.storage)?,
-            &self.ado_type.load(deps.storage)?,
-        )?;
-        Ok(Response::default()
-            .add_attribute("action", "alter_module")
-            .add_attribute("module_idx", module_idx))
-    }
-
-    /// A wrapper for `fn deregister_module`.
-    pub(crate) fn execute_deregister_module(
-        &self,
-        deps: DepsMut,
-        info: MessageInfo,
-        module_idx: Uint64,
-    ) -> Result<Response, ContractError> {
-        let addr = info.sender.as_str();
-        require(
-            self.is_owner_or_operator(deps.storage, addr)?,
-            ContractError::Unauthorized {},
-        )?;
-        self.deregister_module(deps.storage, module_idx)?;
-        Ok(Response::default()
-            .add_attribute("action", "deregister_module")
-            .add_attribute("module_idx", module_idx))
     }
 
     /// Registers a module
@@ -243,11 +180,7 @@ impl<'a> ADOContract<'a> {
         }*/
 
     /// Validates all modules.
-    pub fn validate_modules(
-        &self,
-        modules: &[Module],
-        ado_type: &str,
-    ) -> Result<(), ContractError> {
+    fn validate_modules(&self, modules: &[Module], ado_type: &str) -> Result<(), ContractError> {
         for module in modules {
             module.validate(modules, ado_type)?;
         }
