+++ packages/ado_base/src/state.rs
@@ -1,23 +1,23 @@
 #[cfg(feature = "modules")]
 use common::ado_base::modules::Module;
-use common::{ado_base::QueryMsg, error::ContractError, parse_message};
+use common::{error::ContractError, parse_message};
 use cosmwasm_std::{Addr, Binary, Storage};
 use cw_storage_plus::{Item, Map};
+use serde::de::DeserializeOwned;
 #[cfg(feature = "withdraw")]
 use terraswap::asset::AssetInfo;
 
-//TODO: Make as many of these as possible pub(crate)
 pub struct ADOContract<'a> {
     pub(crate) owner: Item<'a, Addr>,
     pub(crate) operators: Map<'a, &'a str, bool>,
-    pub ado_type: Item<'a, String>,
+    pub(crate) ado_type: Item<'a, String>,
     pub(crate) mission_contract: Item<'a, Addr>,
     #[cfg(feature = "primitive")]
-    pub primitive_contract: Item<'a, Addr>,
+    pub(crate) primitive_contract: Item<'a, Addr>,
     #[cfg(feature = "modules")]
-    pub module_info: Map<'a, &'a str, Module>,
+    pub(crate) module_info: Map<'a, &'a str, Module>,
     #[cfg(feature = "modules")]
-    pub module_idx: Item<'a, u64>,
+    pub(crate) module_idx: Item<'a, u64>,
     #[cfg(feature = "withdraw")]
     pub withdrawable_tokens: Map<'a, &'a str, AssetInfo>,
 }
@@ -42,33 +42,6 @@ impl<'a> Default for ADOContract<'a> {
 }
 
 impl<'a> ADOContract<'a> {
-    /// Helper function to query if a given address is a operator.
-    ///
-    /// Returns a boolean value indicating if the given address is a operator.
-    pub fn is_operator(&self, storage: &dyn Storage, addr: &str) -> bool {
-        self.operators.has(storage, addr)
-    }
-
-    /// Helper function to query if a given address is the current contract owner.
-    ///
-    /// Returns a boolean value indicating if the given address is the contract owner.
-    pub fn is_contract_owner(
-        &self,
-        storage: &dyn Storage,
-        addr: &str,
-    ) -> Result<bool, ContractError> {
-        let owner = self.owner.load(storage)?;
-        Ok(addr == owner)
-    }
-
-    pub fn is_owner_or_operator(
-        &self,
-        storage: &dyn Storage,
-        addr: &str,
-    ) -> Result<bool, ContractError> {
-        Ok(self.is_contract_owner(storage, addr)? || self.is_operator(storage, addr))
-    }
-
     pub fn get_mission_contract(
         &self,
         storage: &dyn Storage,
@@ -87,8 +60,8 @@ impl<'a> ADOContract<'a> {
         Ok(())
     }
 
-    pub(crate) fn is_nested(&self, data: &Option<Binary>) -> bool {
-        let res: Result<QueryMsg, ContractError> = parse_message(data);
+    pub(crate) fn is_nested<T: DeserializeOwned>(&self, data: &Option<Binary>) -> bool {
+        let res: Result<T, ContractError> = parse_message(data);
         res.is_ok()
     }
 }
